+++ src/main/java/br/com/fiap/onlineshop/OnlineShop/dto/ProductUpdateData.java
@@ -10,7 +10,7 @@ import java.math.BigDecimal;
 public record ProductUpdateData(
         String name,
         @DecimalMin(value = "0.0", inclusive = false, message = "O preço deve ser maior que 0")
-        Float price,
+        BigDecimal price,
         String category,
         String description,
         Boolean ativo
